# server/src/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Car API",
        "description": "API was created for the project to car manage and their service requests. The endpoints scheduled below were implemented to create the application as part of the \"Tworzenie serwisów web 2.0\" course.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/register": {
            "post": {
                "summary": "Registers a new user",
                "description": "Registers a new user with the provided information",
                "requestBody": {
                    "description": "User information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    },
                                    "emailAddress": {
                                        "type": "string"
                                    },
                                    "phoneNumber": {
                                        "type": "number"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "firstName",
                                    "lastName",
                                    "emailAddress",
                                    "phoneNumber",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User has been registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User with given email address already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Users"
                ]
            }
        },
        "/login": {
            "post": {
                "summary": "Logs a user in",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginCredentials"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User successfully logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Users"
                ]
            }
        },
        "/session": {
            "get": {
                "summary": "Get user session information",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the user session information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authorization token is required"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "tags": [
                    "Users"
                ]
            }
        },
        "/logout": {
            "post": {
                "summary": "Logout the currently logged in user.",
                "responses": {
                    "200": {
                        "description": "User has been logged out successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "User has been logged out successfully."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refresh token missing or invalid."
                    },
                    "404": {
                        "description": "Refresh token not found."
                    },
                    "500": {
                        "description": "Server error."
                    }
                },
                "tags": [
                    "Users"
                ]
            }
        },
        "/change-password": {
            "post": {
                "summary": "Change user password",
                "description": "Changes the user's password given the old password and new password. Requires a valid refresh token to be sent in the request cookies.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "oldPassword": {
                                        "type": "string",
                                        "description": "The user's old password."
                                    },
                                    "newPassword": {
                                        "type": "string",
                                        "description": "The user's new password."
                                    }
                                },
                                "required": [
                                    "oldPassword",
                                    "newPassword"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The password has been changed successfully."
                    },
                    "401": {
                        "description": "Unauthorized - the request does not contain a valid refresh token or user ID."
                    },
                    "404": {
                        "description": "The user with the given ID does not exist."
                    },
                    "500": {
                        "description": "Server error - an unexpected error occurred on the server."
                    }
                },
                "tags": [
                    "Users"
                ]
            }
        }
    },
...
}